+++ src/Migrate/Migrate/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+* Added validation for Data.Replication
+  - Added validation to protect virtual machines in `New-AzMigrateLocalServerReplication`
+
 ## Version 2.7.0
 * Updated Data.Replication to newer API version
   - Updated Data.Replication to point to stable API version 2024-09-01